(fix:connectorTree) path navigation fn
